requirements.txt
opencv-python==4.8.0.76
mediapipe==0.10.5
tensorflow==2.15.0
numpy==1.25.0
requests==2.31.0
pillow==10.0.0
